Split typechecker program checking
