+++ src/typechecker/mod.rs
@@ -16,6 +16,7 @@ mod environment;
 mod expressions;
 mod monomorphize;
 mod patterns;
+mod program_checking;
 mod resolve;
 mod resolver_validation;
 mod statements;
@@ -109,331 +110,6 @@ impl TypeChecker {
             resolver_missing_behavior_impl_refs: HashSet::new(),
             resolver_missing_behavior_required_refs: HashSet::new(),
         }
-    }
-
-    /// Type-check a program and produce a TypedProgram.
-    pub fn check_program(
-        &mut self,
-        program: &ast::Program,
-    ) -> Result<TypedProgram, Vec<Diagnostic>> {
-        // Phase 1: Collect type definitions and function signatures
-        self.collect_declarations(&program.declarations);
-        self.validate_collected_declaration_semantics(&program.declarations, None);
-        self.check_program_after_collection(program)
-    }
-
-    fn check_program_after_collection(
-        &mut self,
-        program: &ast::Program,
-    ) -> Result<TypedProgram, Vec<Diagnostic>> {
-        // Phase 2: Check function bodies and produce typed AST
-        let mut functions = Vec::new();
-        let mut types = Vec::new();
-        let mut globals = Vec::new();
-        let mut entry_point = None;
-
-        for decl in &program.declarations {
-            match decl {
-                Declaration::Function {
-                    name,
-                    type_params,
-                    params,
-                    return_type,
-                    body,
-                    span,
-                    ..
-                } => {
-                    if !type_params.is_empty() {
-                        continue;
-                    }
-                    if name == "main" {
-                        entry_point = Some(name.clone());
-                    }
-                    match self.check_function(name, params, return_type, body, span) {
-                        Ok(func) => functions.push(func),
-                        Err(d) => self.diagnostics.push(d),
-                    }
-                }
-                Declaration::Method {
-                    type_name,
-                    method_name,
-                    type_params,
-                    params,
-                    return_type,
-                    body,
-                    span,
-                    ..
-                } => {
-                    if !type_params.is_empty() {
-                        continue;
-                    }
-                    let full_name = Self::method_key(type_name, method_name);
-                    // Set Self type for method body
-                    self.current_self_type =
-                        Some(self.resolve_type(&AstType::Named(type_name.clone())));
-                    match self.check_function(&full_name, params, return_type, body, span) {
-                        Ok(func) => functions.push(func),
-                        Err(d) => self.diagnostics.push(d),
-                    }
-                    self.current_self_type = None;
-                }
-                Declaration::Struct {
-                    name,
-                    type_params,
-                    fields,
-                    span,
-                    ..
-                } => {
-                    if !type_params.is_empty() {
-                        continue;
-                    }
-                    let resolved_fields: Vec<(String, Type)> = fields
-                        .iter()
-                        .map(|f| (f.name.clone(), self.resolve_type(&f.ty)))
-                        .collect();
-                    types.push(TypedTypeDef {
-                        name: name.clone(),
-                        kind: TypeDefKind::Struct {
-                            fields: resolved_fields,
-                        },
-                        methods: Vec::new(),
-                        span: *span,
-                    });
-                }
-                Declaration::Enum {
-                    name,
-                    type_params,
-                    variants,
-                    span,
-                    ..
-                } => {
-                    if !type_params.is_empty() {
-                        continue;
-                    }
-                    let typed_variants: Vec<TypedVariant> = variants
-                        .iter()
-                        .enumerate()
-                        .map(|(i, v)| TypedVariant {
-                            name: v.name.clone(),
-                            tag: i as u32,
-                            payload: v
-                                .payload
-                                .as_ref()
-                                .map(|ty| vec![("payload".to_string(), self.resolve_type(ty))]),
-                        })
-                        .collect();
-                    types.push(TypedTypeDef {
-                        name: name.clone(),
-                        kind: TypeDefKind::Enum {
-                            variants: typed_variants,
-                        },
-                        methods: Vec::new(),
-                        span: *span,
-                    });
-                }
-                Declaration::TopLevelExpr { expr, span } => {
-                    // Top-level expressions like main() call
-                    match self.check_expr(expr) {
-                        Ok(typed_expr) => {
-                            globals.push(TypedGlobal {
-                                name: "__top_level__".into(),
-                                ty: typed_expr.ty.clone(),
-                                value: typed_expr,
-                                mutable: false,
-                                span: *span,
-                            });
-                        }
-                        Err(d) => self.diagnostics.push(d),
-                    }
-                }
-                Declaration::Import { .. } => {
-                    // Imports are handled by the module system, not the typechecker
-                }
-                Declaration::Behavior { .. } => {}
-                Declaration::ImplBlock {
-                    type_name,
-                    behavior,
-                    behavior_type_args,
-                    methods,
-                    ..
-                } => {
-                    for method in methods {
-                        if let Declaration::Function {
-                            name,
-                            type_params,
-                            params,
-                            return_type,
-                            body,
-                            span,
-                            ..
-                        } = method
-                        {
-                            if !type_params.is_empty() {
-                                continue;
-                            }
-                            let full_name = Self::method_key(type_name, name);
-                            self.current_self_type =
-                                Some(self.resolve_type(&AstType::Named(type_name.clone())));
-                            match self.check_function(&full_name, params, return_type, body, span) {
-                                Ok(func) => functions.push(func),
-                                Err(d) => self.diagnostics.push(d),
-                            }
-                            self.current_self_type = None;
-                        }
-                    }
-
-                    if let Some(behavior) = behavior {
-                        for default in self.behavior_default_methods_for_impl(
-                            type_name,
-                            behavior,
-                            behavior_type_args,
-                            methods,
-                        ) {
-                            let full_name = Self::method_key(type_name, &default.name);
-                            self.current_self_type =
-                                Some(self.resolve_type(&AstType::Named(type_name.clone())));
-                            match self.check_function(
-                                &full_name,
-                                &default.params,
-                                &default.return_type,
-                                &default.body,
-                                &default.span,
-                            ) {
-                                Ok(func) => functions.push(func),
-                                Err(d) => self.diagnostics.push(d),
-                            }
-                            self.current_self_type = None;
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-
-        let errors: Vec<_> = self
-            .diagnostics
-            .iter()
-            .filter(|d| d.is_error())
-            .cloned()
-            .collect();
-        if !errors.is_empty() {
-            return Err(errors);
-        }
-
-        functions.append(&mut self.specialized_functions);
-        types.append(&mut self.specialized_types);
-
-        Ok(TypedProgram {
-            functions,
-            types,
-            globals,
-            entry_point,
-        })
-    }
-
-    pub fn check_program_with_symbols(
-        &mut self,
-        program: &ast::Program,
-        symbols: &SymbolTable,
-    ) -> Result<TypedProgram, Vec<Diagnostic>> {
-        self.validate_resolver_symbols(program, symbols);
-        if self.diagnostics.iter().any(|diag| diag.is_error()) {
-            return Err(self
-                .diagnostics
-                .iter()
-                .filter(|diag| diag.is_error())
-                .cloned()
-                .collect());
-        }
-        self.collect_resolver_imports(symbols);
-        self.collect_declarations_with_symbols(&program.declarations, symbols);
-        self.check_program_after_collection(program)
-    }
-
-    pub fn check_module_graph_entry(
-        &mut self,
-        graph: &ResolvedModuleGraph,
-    ) -> Result<TypedProgram, Vec<Diagnostic>> {
-        let Some(entry) = graph.module(graph.entry) else {
-            self.diagnostics.push(Diagnostic::error(
-                "E0232",
-                format!("module graph missing entry module {:?}", graph.entry),
-                Span::dummy(),
-            ));
-            return Err(self
-                .diagnostics
-                .iter()
-                .filter(|diag| diag.is_error())
-                .cloned()
-                .collect());
-        };
-
-        let mut modules = graph.modules().values().collect::<Vec<_>>();
-        modules.sort_by_key(|module| module.info.id.0);
-
-        let mut dependency_programs = Vec::new();
-        for module in modules {
-            if module.info.id == graph.entry {
-                continue;
-            }
-
-            let mut checker = TypeChecker::new();
-            match checker.check_module_graph_module(graph, module) {
-                Ok(typed) => dependency_programs.push(typed),
-                Err(diags) => self.diagnostics.extend(diags),
-            }
-        }
-
-        if self.diagnostics.iter().any(|diag| diag.is_error()) {
-            return Err(self
-                .diagnostics
-                .iter()
-                .filter(|diag| diag.is_error())
-                .cloned()
-                .collect());
-        }
-
-        let mut typed = self.check_module_graph_module(graph, entry)?;
-        for mut dependency in dependency_programs {
-            typed.functions.append(&mut dependency.functions);
-            typed.types.append(&mut dependency.types);
-            typed.globals.append(&mut dependency.globals);
-        }
-        Ok(typed)
-    }
-
-    fn check_module_graph_module(
-        &mut self,
-        graph: &ResolvedModuleGraph,
-        module: &ResolvedModule,
-    ) -> Result<TypedProgram, Vec<Diagnostic>> {
-        self.validate_resolver_symbols(&module.program, &module.symbols);
-        if self.diagnostics.iter().any(|diag| diag.is_error()) {
-            return Err(self
-                .diagnostics
-                .iter()
-                .filter(|diag| diag.is_error())
-                .cloned()
-                .collect());
-        }
-
-        self.collect_module_graph_imports(graph, module);
-        if self.diagnostics.iter().any(|diag| diag.is_error()) {
-            return Err(self
-                .diagnostics
-                .iter()
-                .filter(|diag| diag.is_error())
-                .cloned()
-                .collect());
-        }
-
-        self.collect_declarations_with_symbols(&module.program.declarations, &module.symbols);
-        self.check_program_after_collection(&module.program)
-    }
-
-    /// Get all diagnostics (errors + warnings).
-    pub fn diagnostics(&self) -> &[Diagnostic] {
-        &self.diagnostics
     }
 
     // ── Phase 1: Collect ──────────────────────────────────────────
